fix static

+++ client/components/features/admin/AdminSidebar.jsx
@@ -94,9 +94,11 @@ export default function AdminSidebar({ sidebarOpen, setSidebarOpen, currentPath 
     ]
 
     const handleLogout = () => {
-        localStorage.clear()
-        document.cookie = 'authToken=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT'
-        document.cookie = 'roles=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT'
+        if (typeof window !== 'undefined') {
+            localStorage.clear()
+            document.cookie = 'authToken=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT'
+            document.cookie = 'roles=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT'
+        }
         router.push('/signin')
     }
